Separate main impl functionality into separate files and fix a small mistake
